Moving AbstractTrackConfiguration to Custom Gui
