+++ org.mwc.debrief.lite/src/main/java/org/mwc/debrief/lite/gui/custom/JSelectTrackModel.java
@@ -5,8 +5,6 @@ import java.beans.PropertyChangeListener;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.mwc.debrief.lite.graph.AbstractTrackConfiguration;
-
 import Debrief.Wrappers.TrackWrapper;
 import MWC.GUI.Layer;
 import MWC.GUI.Layers;
